+++ src/TableViewerBlazor/Options/TvStyleOption.cs
@@ -1,0 +1,9 @@
+namespace TableViewerBlazor.Options;
+
+public class TvStyleOption
+{
+    public bool Dense { get; set; } = true;
+    public bool Hover { get; set; } = false;
+    public bool Striped { get; set; } = false;
+    public bool Bordered { get; set; } = true;
+}
